requirements.txt
aiogram==3.13.1
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
python-docx==1.1.2
pdfplumber==0.11.4
deep-translator==1.11.4